requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-filter==24.3
drf-spectacular==0.27.2
django-cors-headers==4.5.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
PyYAML==6.0.3
jsonschema==4.25.1
